move to conda.testing.fixtures
